Abstracted RobotComponents into seperate file
